+++ src/main/java/me/zeroeightsix/kami/module/modules/movement/TimerSpeed.java
@@ -4,7 +4,8 @@ import me.zeroeightsix.kami.module.Module;
 import me.zeroeightsix.kami.module.Module.Category;
 import me.zeroeightsix.kami.setting.Setting;
 import me.zeroeightsix.kami.setting.Settings;
-import me.zeroeightsix.kami.util.InfoCalculator;
+
+import static me.zeroeightsix.kami.util.MathsUtils.round;
 
 /**
  * @author TBM
@@ -24,7 +25,7 @@ public class TimerSpeed extends Module {
     private Setting<Float> fastSpeed = register(Settings.floatBuilder("Fast Speed").withMinimum(1.0F).withMaximum(10.0F).withValue(5.0F).build());
 
     public static String returnGui() {
-        return "" + InfoCalculator.round(curSpeed, 2);
+        return "" + round(curSpeed, 2);
     }
 
     public void onUpdate() {
