+++ src/main/java/me/zeroeightsix/kami/util/MathsUtils.java
@@ -1,5 +1,10 @@
 package me.zeroeightsix.kami.util;
 
+import net.minecraft.client.Minecraft;
+import net.minecraft.util.math.MathHelper;
+
+import java.text.DecimalFormat;
+
 /**
  * Created by Dewy on the 17th of April, 2020
  */
@@ -16,4 +21,13 @@ public class MathsUtils {
 
         return angleIn;
     }
+
+    public static double round(double value, int places) {
+        double scale = Math.pow(10, places);
+        return Math.round(value * scale) / scale;
+    }
+
+    public static boolean isNumberEven(int i) { return (i & 1) == 0; }
+
+    public static int reverseNumber(int num, int min, int max) { return (max + min) - num; }
 }
